Move bootstrap package from internal to pkg
