+++ pkg/bootstrap/git/gogit/testdata/private.key
